fix(convert): align Market Data contracts
Require a documented pair filter, validate signed asset precision receive windows, expose int64 fraction values, and add deterministic request/model coverage for both current Convert Market Data endpoints.

+++ Binance.Api/Convert/Responses/BinanceConvertAsset.cs
@@ -13,5 +13,5 @@ public record BinanceConvertAsset
     /// <summary>
     /// Fraction
     /// </summary>
-    public int Fraction { get; set; }
+    public long Fraction { get; set; }
 }
